+++ README.md
@@ -2,6 +2,48 @@
 ## 요구사항
 
 ### 기본
+
+#### 스프린트 미션 4
+
+###### 인증
+- [x] User 스키마를 작성해 주세요(id, email, nickname, image, password, createdAt, updatedAt 필드를 가집니다.)
+- [ ] 회원가입 API를 만들어 주세요.(email, nickname, password 를 입력하여 회원가입을 진행합니다. password는 해싱해 저장합니다.)
+- [ ] 토큰 기반 인증: 로그인에 성공하면 Access Token을 발급하는 기능을 구현합니다.
+
+###### 상품 기능 인가
+- [ ] 로그인한 유저만 상품을 등록할 수 있습니다.
+- [ ] 상품을 등록한 유저만 해당 상품의 정보를 수정하거나 삭제할 수 있습니다.
+
+###### 게시글 기능 인가
+- [ ] 로그인한 유저만 상품을 등록할 수 있습니다.
+- [ ] 게시글을 등록한 유저만 해당 게시글을 수정하거나 삭제할 수 있습니다.
+
+###### 댓글 기능 인가
+- [ ] 로그인한 유저만 상품에 댓글을 등록할 수 있습니다.
+- [ ] 로그인한 유저만 게시글에 댓글을 등록할 수 있습니다.
+- [ ] 댓글을 등록한 유저만 해당 댓글을 수정하거나 삭제할 수 있습니다.
+
+###### 유저 정보
+- [ ] 유저가 자신의 정보를 조회하는 기능을 구현합니다.
+- [ ] 유저가 자신의 정보를 수정할 수 있는 기능을 구현합니다.
+- [ ] 유저가 자신의 비밀번호를 변경할 수 있는 기능을 구현합니다.
+- [ ] 유저가 자신이 등록한 상품의 목록을 조회하는 기능을 구현합니다.
+- [ ] 유저의 비밀번호는 리스폰스로 노출하지 않습니다.
+
+
+### 심화
+
+###### 인증
+- [ ] 토큰 기반 인증: Refresh Token으로 토큰을 갱신하는 기능을 구현합니다.
+
+###### 좋아요 기능
+- [ ] 로그인한 유저는 상품에 '좋아요'와 '좋아요 취소'를 할 수 있습니다.
+- [ ] 로그인한 유저는 게시글에 '좋아요'와 '좋아요 취소'를 할 수 있습니다.
+- [ ] 상품 또는 게시글을 조회할 때, 유저가 '좋아요'를 누른 항목인지 확인할 수 있도록 isLiked와 같은 불린형 필드를 리스폰스 객체에 포함시켜 리스폰스해 주세요.
+- [ ] 유저가 '좋아요'를 표시한 상품의 목록을 조회하는 기능을 구현합니다.
+
+## 스프린트 3
+
 - [x] PostgreSQL를 이용해 주세요.
 - [x] 데이터 모델 간의 관계를 고려하여 onDelete를 설정해 주세요.
 - [x] 데이터베이스 시딩 코드를 작성해 주세요.
@@ -46,10 +88,6 @@
 - [x] .env 파일에 환경 변수를 설정해 주세요.
 - [x] CORS를 설정해 주세요.
 - [ ] render.com으로 배포해 주세요.
-
-### 심화
-- [ ] 심화 항목 1
-- [ ] 심화 항목 2
 
 ## 주요 변경사항
 -
