+++ prisma/schema.prisma
@@ -7,9 +7,15 @@ datasource db {
   url      = env("DATABASE_URL")
 }
 
-// model User {
-//   id          String @id @default(())
-// }
+model User {
+  id        Int      @id @default(autoincrement())
+  email     String   @unique
+  nickname  String
+  image     String?
+  password  String
+  createdAt DateTime @default(now())
+  updatedAt DateTime @updatedAt
+}
 
 model Product {
   id          String    @id @default(uuid())
